+++ pkg/util/buckets/buckets.go
@@ -21,7 +21,7 @@ type Bucketable interface {
 
 type WorkerFunc func(<-chan struct{}, string)
 
-type monitor[E Bucketable] struct {
+type cache[E Bucketable] struct {
 	baseLog *logrus.Entry
 
 	bucketCount int
@@ -43,8 +43,8 @@ type BucketWorker[E Bucketable] interface {
 	UpsertDoc(E)
 }
 
-func NewBucketWorker[E Bucketable](log *logrus.Entry, worker WorkerFunc, mu *sync.RWMutex) *monitor[E] {
-	return &monitor[E]{
+func NewBucketWorker[E Bucketable](log *logrus.Entry, worker WorkerFunc, mu *sync.RWMutex) *cache[E] {
+	return &cache[E]{
 		baseLog: log,
 
 		worker: worker,
@@ -58,11 +58,11 @@ func NewBucketWorker[E Bucketable](log *logrus.Entry, worker WorkerFunc, mu *syn
 }
 
 // Return the size of the document cache. Caller must hold mon.mu.
-func (mon *monitor[E]) Size() int {
+func (mon *cache[E]) Size() int {
 	return len(mon.docs)
 }
 
-func (mon *monitor[E]) Doc(id string) (r E, ok bool) {
+func (mon *cache[E]) Doc(id string) (r E, ok bool) {
 	mon.mu.RLock()
 	defer mon.mu.RUnlock()
 	id = strings.ToLower(id)
@@ -74,7 +74,7 @@ func (mon *monitor[E]) Doc(id string) (r E, ok bool) {
 	return v.doc, true
 }
 
-func (mon *monitor[E]) SetBuckets(buckets []int) {
+func (mon *cache[E]) SetBuckets(buckets []int) {
 	mon.mu.Lock()
 	defer mon.mu.Unlock()
 	oldBuckets := mon.buckets
@@ -88,6 +88,82 @@ func (mon *monitor[E]) SetBuckets(buckets []int) {
 		mon.baseLog.Printf("servicing %d buckets", len(mon.buckets))
 		for _, v := range mon.docs {
 			mon.FixDoc(v.doc)
+		}
+	}
+}
+
+type cacheDoc[E Bucketable] struct {
+	doc  E
+	stop chan<- struct{}
+}
+
+// deleteDoc deletes the given document from mon.docs, signalling the associated
+// monitoring goroutine to stop if it exists.  Caller must hold mon.mu.Lock.
+func (mon *cache[E]) DeleteDoc(doc E) {
+	id := strings.ToLower(doc.GetID())
+	v := mon.docs[id]
+
+	if v != nil {
+		if v.stop != nil {
+			mon.baseLog.Debugf("deleting doc, closing worker for %s", doc.GetID())
+			close(mon.docs[id].stop)
+		}
+
+		delete(mon.docs, id)
+	}
+}
+
+// upsertDoc inserts or updates the given document into mon.docs, starting an
+// associated monitoring goroutine if the document is in a bucket owned by us.
+// Caller must hold mon.mu.Lock.
+func (mon *cache[E]) UpsertDoc(doc E) {
+	id := strings.ToLower(doc.GetID())
+	v := mon.docs[id]
+
+	if v == nil {
+		v = &cacheDoc[E]{}
+		mon.docs[id] = v
+	}
+
+	v.doc = doc
+	mon.FixDoc(doc)
+}
+
+// fixDoc ensures that there is a monitoring goroutine for the given document
+// if it is in a bucket owned by us.  Caller must hold mon.mu.Lock.
+func (mon *cache[E]) FixDoc(doc E) {
+	id := strings.ToLower(doc.GetID())
+	v := mon.docs[id]
+
+	var ours bool
+	// getBucket() with -1 is served by all
+	if v.doc.GetBucket() > -1 {
+		_, ours = mon.buckets[v.doc.GetBucket()]
+	} else {
+		ours = true
+	}
+
+	if !ours && v.stop != nil {
+		mon.baseLog.Debugf("we no longer own cluster, closing worker for %s", doc.GetID())
+		close(v.stop)
+		v.stop = nil
+	} else if ours && v.stop == nil {
+		ch := make(chan struct{})
+		v.stop = ch
+
+		mon.baseLog.Debugf("spawning worker for %s", doc.GetID())
+		mon.worker(ch, doc.GetKey())
+	}
+}
+
+// Stop stops all workers.
+func (mon *cache[E]) Stop() {
+	mon.mu.Lock()
+	defer mon.mu.Unlock()
+	for _, v := range mon.docs {
+		if v.stop != nil {
+			close(v.stop)
+			v.stop = nil
 		}
 	}
 }
